Major Refactor
Added unit tests
Local Persistence for debugging with the Alexa Skills Kit (ASK) Toolkit VSCode extension
Refactored index.js and split core code into separate modules.
