+++ skill/lambda/package.json
@@ -10,10 +10,15 @@
   "license": "ISC",
   "dependencies": {
     "ask-sdk-core": "^2.6.0",
-    "ask-sdk-model": "^1.18.0",
-    "aws-sdk": "^2.326.0",
+    "ask-sdk-model": "^1.29.0",
     "ask-sdk-s3-persistence-adapter": "^2.0.0",
+    "aws-sdk": "^2.326.0",
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.28"
+  },
+  "devDependencies": {
+    "ask-sdk-local-debug": "^1.0.0",
+    "jasmine-node": "3.0.0",
+    "js-sha512": "^0.8.0"
   }
 }
