+++ skill/lambda/spec/profiles/ss-one-hour-greater-no-water-50-health.json
@@ -1,0 +1,14 @@
+{
+    "cactus": {
+      "waterLevel": 0,
+      "healthLevel": 49,
+      "dayOfBirth": 1588890484157,
+      "daysAlive": 0,
+      "blindState": "closed",
+      "isCold": false,
+      "lastUpdated": 1588890484157,
+      "latestInteraction": 1588895633166,
+      "name": "Sharpie Sharp"
+    },
+    "timeZone": "America/New_York"
+  }
